sửa services để khớp với API mới

+++ src/services/khoa/khoaRequest.js
@@ -7,9 +7,13 @@ export const khoaRequest = {
       let url = `/Khoas`;
       let res = await axiosInstance.post(url, data);
       console.log("Kết quả trả về khi thêm khoa: ", res);
-      dispatch(add(res.data));
+      if (res.status === 201) {
+        dispatch(add(res.data.khoa));
+      }
+      return { status: res.status, data: res.data };
     } catch (err) {
       console.log("Thêm khoa bị lỗi: ", err);
+      return { status: err.status, data: err.data };
     }
   },
   getAll: async function (dispatch) {
@@ -17,9 +21,13 @@ export const khoaRequest = {
       let url = `/Khoas`;
       let res = await axiosInstance.get(url);
       console.log("Kết quả trả về khi lấy danh sách khoa: ", res);
-      dispatch(updateAll(res.data));
+      if (res.status === 200) {
+        dispatch(updateAll(res.data.khoas));
+      }
+      return { status: res.status, data: res.data };
     } catch (err) {
       console.log("Lấy danh sách khoa bị lỗi: ", err);
+      return { status: err.status, data: err.data };
     }
   },
   delete: async function (id, dispatch) {
@@ -30,13 +38,18 @@ export const khoaRequest = {
       }
 
       if (isDelete) {
-        dispatch(deleteOne(id));
         let url = `/Khoas/${id}`;
         let res = await axiosInstance.delete(url);
+        if(res.status === 200)  {
+          dispatch(deleteOne(id));
+        }
         console.log("Kết quả trả về khi xóa khoa: ", res);
+        return { status: res.status, data: res.data };
       }
+      return {};
     } catch (err) {
       console.log("Lỗi khi xóa khoa có id = ", id);
+      return { status: err.status, data: err.data };
     }
   },
   update: async function (data, dispatch) {
@@ -49,9 +62,12 @@ export const khoaRequest = {
         let url = `/Khoas/${data.id}`;
         let res = await axiosInstance.put(url);
         console.log("Kết quả trả về khi sửa khoa: ", res);
+        return { status: res.status, data: res.data };
       }
+      return {};
     } catch (err) {
       console.log("Lỗi khi sửa khoa có id = ", data.id);
+      return { status: err.status, data: err.data };
     }
   },
   getAllGiangVienByKhoa: async function (id) {
@@ -63,10 +79,10 @@ export const khoaRequest = {
         id,
         res.data
       );
-      return res.data;
+      return { status: res.status, data: res.data };
     } catch (err) {
       console.log("Lỗi khi lấy danh sách giảng viên có khoaId =", id, err);
-      return err;
+      return { status: err.status, data: err.data };
     }
   },
   getAllSinhVienByKhoa: async function (id) {
@@ -78,10 +94,10 @@ export const khoaRequest = {
         id,
         res.data
       );
-      return res.data;
+      return { status: res.status, data: res.data };
     } catch (err) {
       console.log("Lỗi khi lấy danh sách sinh viên có khoaId =", id, err);
-      return err;
+      return { status: err.status, data: err.data };
     }
   },
 };
